+++ src/frontend-cdoc-commons/components/cdoc-audioplayer/cdoc-audioplayer.component.ts
@@ -18,6 +18,13 @@ import {AbstractInlineComponent} from '../../../angular-commons/components/inlin
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class CommonDocAudioplayerComponent extends AbstractInlineComponent {
+    maxWidth = 600;
+    maxHeight = 800;
+    maxFullWidth = 1200;
+    maxFullHeight = 1200;
+    flgIsPlaying = false;
+    playerId;
+
     public contentUtils: CommonDocContentUtils;
     listItem: CommonItemData = {
         currentRecord: undefined,
@@ -42,6 +49,18 @@ export class CommonDocAudioplayerComponent extends AbstractInlineComponent {
     @Input()
     public styleClass: 'picture-small';
 
+    @Input()
+    public playerIdPrefix ? = 'player';
+
+    @Input()
+    public autoplay?: boolean = false;
+
+    @Output()
+    public playerStarted: EventEmitter<CommonDocRecord> = new EventEmitter();
+
+    @Output()
+    public playerStopped: EventEmitter<CommonDocRecord> = new EventEmitter();
+
     @Output()
     public show: EventEmitter<CommonDocRecord> = new EventEmitter();
 
@@ -55,18 +74,54 @@ export class CommonDocAudioplayerComponent extends AbstractInlineComponent {
         return false;
     }
 
+    onPlayerStarted() {
+        this.startPlaying();
+        this.flgIsPlaying = true;
+        this.playerStarted.emit(this.record);
+    }
+
+    onPlayerEnded() {
+        this.flgIsPlaying = false;
+        this.playerStopped.emit(this.record);
+    }
+
+    onPlayerPaused() {
+        this.flgIsPlaying = false;
+    }
+
+    protected updateData(): void {
+        if (window) {
+            this.maxWidth = Math.min(600, window.innerWidth - 100);
+            this.maxHeight = Math.min(800, window.innerHeight - 80);
+            this.maxFullWidth = Math.min(1200, window.innerWidth - 50);
+            this.maxFullHeight = Math.min(1200, window.innerHeight - 80);
+        }
+        this.contentUtils.updateItemData(this.listItem, this.record, 'default');
+        this.playerId = this.playerIdPrefix + '_' + this.listItem.currentRecord.id;
+        this.cd.markForCheck();
+        this.startAutoPlayPlayer();
+    }
+
     startPlaying() {
         const audios = document.querySelectorAll('audio');
         for (let i = 0; i < audios.length; i++) {
             const audio = audios[i];
-            if (audio.getAttribute('mediaid') !== this.listItem.currentRecord.id) {
+            if (audio.getAttribute('mediaid') !== this.playerId) {
                 audio.pause();
             }
         }
     }
 
-    protected updateData(): void {
-        this.contentUtils.updateItemData(this.listItem, this.record, 'default');
-        this.cd.markForCheck();
+    startAutoPlayPlayer(): void {
+        if (this.autoplay) {
+            const audios = document.querySelectorAll('audio');
+            for (let i = 0; i < audios.length; i++) {
+                const audio = audios[i];
+                if (audio.getAttribute('mediaid') === this.playerId) {
+                    audio.play();
+                }
+            }
+        }
     }
 }
+